// lib/database.types.ts
export interface Database {
  public: {
    Tables: {
// ...
      payment_config: {
        Row: {
          id: number
          provider: 'sumup' | 'stripe'
          is_enabled: boolean
          stripe_private_key: string | null
          stripe_public_key: string | null
          sumup_merchant_email: string | null
          sumup_key: string | null
          created_at: string
          updated_at: string
        }
        Insert: {
          id?: never // Generated always as identity
          provider: 'sumup' | 'stripe'
          is_enabled: boolean
          stripe_private_key?: string | null
          stripe_public_key?: string | null
          sumup_merchant_email?: string | null
          sumup_key?: string | null
          created_at?: string
          updated_at?: string
        }
        Update: {
          provider?: 'sumup' | 'stripe'
          is_enabled?: boolean
          stripe_private_key?: string | null
          stripe_public_key?: string | null
          sumup_merchant_email?: string | null
          sumup_key?: string | null
          updated_at?: string
        }
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
}